IFeature adjustments FEATURE_VERSION changed in anticipation of a new feature. uint96 should not overflow if staying in sensible ranges for version numbers.

+++ contracts/features/interfaces/IFeature.sol
@@ -36,5 +36,5 @@ interface IFeature {
     function FEATURE_NAME() external view returns (string memory name);
 
     /// @notice The version of this feature set.
-    function FEATURE_VERSION() external view returns (uint256 version);
+    function FEATURE_VERSION() external view returns (uint96 version);
 }
